rename files, join files
